+++ src/plat/spike/config.cmake
@@ -1,6 +1,7 @@
 #
 # Copyright 2020, DornerWorks
 # Copyright 2020, Data61, CSIRO (ABN 41 687 119 230)
+# Copyright 2021, HENSOLDT Cyber
 #
 # SPDX-License-Identifier: GPL-2.0-only
 #
@@ -27,7 +28,7 @@ if(KernelPlatformSpike)
     endif()
     declare_default_headers(
         TIMER_FREQUENCY 10000000 PLIC_MAX_NUM_INT 0
-        INTERRUPT_CONTROLLER arch/machine/plic.h
+        INTERRUPT_CONTROLLER drivers/irq/riscv_plic_dummy.h
     )
 else()
     unset(KernelPlatformFirstHartID CACHE)
